requirements.txt
annotated-types==0.7.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
httplib2==0.31.0
idna==3.11
instapaper==0.5
Jinja2==3.1.6
MarkupSafe==3.0.3
murmurhash==1.0.15
numpy==2.3.5
oauth2==1.9.0.post1
packaging==25.0
preshed==3.0.12
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
thinc==8.3.10
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1

# Development and linting tools
# Voice command support (STT via pipecat + faster-whisper)
# System requirement: portaudio19-dev  (sudo apt install portaudio19-dev)
pipecat-ai[local,whisper,websockets-base]==0.0.105
websockets==15.0.1

# Remote voice transport (Daily WebRTC + Deepgram STT)
daily-python==0.24.0
deepgram-sdk==6.0.1

# Development and linting tools
black==25.11.0
flake8==7.3.0
isort==7.0.0
mypy==1.18.2
pylint==4.0.3
